requirements.txt
pygame-ce==2.4.1
numpy==1.24.3
dataclasses-json==0.5.9
typing-extensions==4.7.1 